add diverse.py
